GUI and Fixes
Added two fields for the GameManager to take in and their respective start menu objects. I also changed how the Mutation Factor works, it now takes in a percentage and adds or subtracts that from the value (meaning you can never have a negative number and that it can go with any number)
