+++ Boids/Assets/Scripts/Management.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e885329703638c45adb3c8bec0c912b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
