+++ Boids/Assets/Scripts/Management/PauseMenu.cs
@@ -1,0 +1,64 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class PauseMenu : MonoBehaviour
+{
+    public GameObject Pause_Menu;
+    public static bool isPaused;
+
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        Pause_Menu.SetActive(false);
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.Escape))
+        {
+            if (isPaused)
+            {
+                ResumeGame();
+            }
+            else
+            {
+                PauseGame();
+            }
+
+        }
+    }
+
+    public void PauseGame()
+    {
+        Pause_Menu.SetActive(true);
+        Time.timeScale = 0f;
+        isPaused = true;
+    }
+
+    public void ResumeGame()
+    {
+        Pause_Menu.SetActive(false);
+        Time.timeScale = 1f;
+        isPaused = false;
+
+    }
+
+    public void GoToMainMenu()
+    {
+        Time.timeScale = 1f;
+        SceneManager.LoadScene("Start Menu");
+    }
+
+    public void RestartGame()
+    {
+        SceneManager.LoadScene(SceneManager.GetActiveScene().name);
+        Time.timeScale = 1f;
+
+        GetComponent<GameManager>().particleCount = 0;
+        GetComponent<GameManager>().foodCount = 0;
+    }    
+}
