--- Boids/Assets/Scripts/ParticleAmt.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5c73ccd825f85b84797f5354ca572702
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
