--- Boids/Assets/Scripts/PopulationManager.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using TMPro;
-using UnityEngine.UI;
-
-public class PopulationManager : MonoBehaviour
-{
-    
-}
