--- Boids/Assets/Scripts/PopulationManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cc73da6a6ad150146b1f7341b472fcbe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
