+++ Boids/Assets/Scripts/StartMenu/SliderValueGetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 454514ea9a4974348a7fcb4b3307ef30
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
