+++ Boids/Assets/Scripts/Support/FoodAmt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 75107fecae9b3a6418cedabf771d2ddf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
