+++ Boids/Assets/Scripts/Support/ParticleAmt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c73ccd825f85b84797f5354ca572702
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
