+++ Boids/Assets/Scripts/Support/PopulationManager.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using TMPro;
+using UnityEngine.UI;
+
+public class PopulationManager : MonoBehaviour
+{
+    
+}
